GT.M V7.1-002

+++ sr_port/mvalconv.c
@@ -1,6 +1,6 @@
 /****************************************************************
  *								*
- * Copyright (c) 2001-2022 Fidelity National Information	*
+ * Copyright (c) 2001-2023 Fidelity National Information	*
  * Services, Inc. and/or its subsidiaries. All rights reserved.	*
  *								*
  *	This source code contains the intellectual property	*
@@ -313,7 +313,7 @@ int4 mval2i(mval *v)
 		{
 			j = mval2double(v);
 			if (MINNEGINT4 > j)
-				i = MINNEGINT4;
+				i = (int4) MINNEGINT4;
 			else if (MAXPOSINT4 < j)
 				i = MAXPOSINT4;
 			else	/* if ((MINNEGINT4 <= j) && (MAXPOSINT4 >= j)) */
